# Always cache Attribute compatibility and disambiguation rules