--- a/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/ResolvedVariantSet.java
+++ b/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/ivyservice/resolveengine/artifact/ResolvedVariantSet.java
@@ -22,7 +22,7 @@ import org.gradle.api.internal.attributes.AttributesSchemaInternal;
 import org.gradle.api.internal.attributes.ImmutableAttributes;
 
 import javax.annotation.Nullable;
-import java.util.Set;
+import java.util.List;
 
 /**
  * Represents some provider of {@link ResolvedVariant} instances to select from.
@@ -44,7 +44,7 @@ public interface ResolvedVariantSet {
     /**
      * The variants available for artifact selection when variant reselection is not enabled.
      */
-    Set<ResolvedVariant> getVariants();
+    List<ResolvedVariant> getVariants();
 
     /**
      * The provider may have been selected thanks to a different attribute set than the one from