--- a/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/repositories/AbstractArtifactRepository.java
+++ b/platforms/software/dependency-management/src/main/java/org/gradle/api/internal/artifacts/repositories/AbstractArtifactRepository.java
@@ -195,8 +195,13 @@ public abstract class AbstractArtifactRepository implements ArtifactRepositoryIn
 
 
     private static <T> InstantiatingAction<T> createRuleAction(final Instantiator instantiator, final ConfigurableRule<T> rule) {
-        return new InstantiatingAction<>(DefaultConfigurableRules.of(rule), instantiator, (target, throwable) -> {
-            throw UncheckedException.throwAsUncheckedException(throwable);
-        });
+        return new InstantiatingAction<>(
+            DefaultConfigurableRules.of(rule),
+            instantiator,
+            (target, throwable) -> {
+                throw UncheckedException.throwAsUncheckedException(throwable);
+            },
+            false
+        );
     }
 }