--- a/platforms/software/dependency-management/src/test/groovy/org/gradle/internal/resolve/caching/ComponentMetadataRuleExecutorTest.groovy
+++ b/platforms/software/dependency-management/src/test/groovy/org/gradle/internal/resolve/caching/ComponentMetadataRuleExecutorTest.groovy
@@ -206,7 +206,8 @@ class ComponentMetadataRuleExecutorTest extends Specification {
         rule = new InstantiatingAction<>(
             DefaultConfigurableRules.of(DefaultConfigurableRule.of(ruleClass)),
             instantiator,
-            shouldNotFail()
+            shouldNotFail(),
+            false
         )
     }