--- a/platforms/software/dependency-management/src/test/groovy/org/gradle/internal/resolve/caching/ComponentMetadataSupplierRuleExecutorTest.groovy
+++ b/platforms/software/dependency-management/src/test/groovy/org/gradle/internal/resolve/caching/ComponentMetadataSupplierRuleExecutorTest.groovy
@@ -193,7 +193,8 @@ class ComponentMetadataSupplierRuleExecutorTest extends Specification {
         rule = new InstantiatingAction<>(
             DefaultConfigurableRules.of(DefaultConfigurableRule.of(ruleClass)),
             instantiator,
-            shouldNotFail()
+            shouldNotFail(),
+            false
         )
     }