… platforms/core-runtime/base-services/src/main/java/org/gradle/api/internal/ReusableAction.java
… platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/ivyservice/ivyresolve/DefaultMetadataProviderTest.groovy
… platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/repositories/resolver/MavenResolverTest.groovy
… platforms/software/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/transform/DefaultArtifactVariantSelectorFactoryTest.groovy
… platforms/software/dependency-management/src/test/groovy/org/gradle/internal/resolve/caching/CrossBuildCachingRuleExecutorTest.groovy
… subprojects/core/src/main/java/org/gradle/internal/action/InstantiatingAction.java
… subprojects/core/src/test/groovy/org/gradle/internal/action/InstantiatingActionTest.groovy