Compatibility rules especially are called in tight loops in the middle of dependency resolution. We have seen builds that rely on custom rules heavily spend a huge amount of time during resolution instantiating these rules over and over.

This commit updates the rule instantiator to cache the value for the rule so that we do not waste time instantiating the rule over and over again.

We expect these rules to be deterministic. Both AGP and KGP rules were verified to be deterministic and not have side-effects in their constructors.

Fixes [#30110](https://github.com/gradle/gradle/issues/30110)

Running the Android 2000 project extra large benchmark with 10 iterations:

before:

Max: 71.0 seconds
Min: 60.5 seconds
Mean: 63.0 seconds
Median: 62.3 seconds

after:

Max: 63.4 seconds
Min: 53.8 seconds
Mean: 56.4 seconds
Median: 55.4 seconds

This gives an improvement of:

Max: 10.8%
Min: 11.1%
Mean: 10.5%
Median: 11.1%

This is clear looking a the flamegraphs of artifact selection for example
![image](https://github.com/user-attachments/assets/0723cd70-d962-456f-ae38-70c5c3ef359e)
![image](https://github.com/user-attachments/assets/f85c9448-6ab5-4fe9-b13a-9038c8f0cf7d)

Artifact selection went from 10,842 samples to 6,646 samples, a 38.7% improvement

The raw flamegraphs themselves:

[androidStudioSync-8 11-20240809025650+0000-cpu-simplified-flames](https://github.com/user-attachments/assets/ad35a34e-b70b-426a-bed0-2dc54931ccdc)
[androidStudioSync-8 11-20240809022902+0000-cpu-simplified-flames](https://github.com/user-attachments/assets/a93c77c4-ddfe-40f0-ae69-8abfe1ea3792)



### Reviewing cheatsheet

Before merging the PR, comments starting with 
- ❌ ❓**must** be fixed
- 🤔 💅 **should** be fixed
- 💭 **may** be fixed
- 🎉 celebrate happy things