@jvandort 2024-08-13
We will not always cache these since we want to enforce statelessness of the rules. 

Instad, we will:
1. Use https://github.com/gradle/gradle/blob/9239189ef8d307b43b765f5b9637f509d8f5a436/platforms/core-configuration/model-core/src/main/java/org/gradle/internal/instantiation/InstantiationScheme.java#L38 to improve instantiation performance
2. Cache results of each rule call
3. Hoist up attribute schema caches so we deduplicate caches between projects